backend/project/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
packaging==24.2
PyJWT==2.10.1
pyotp==2.9.0
sqlparse==0.5.3
